requirements.txt
google-adk[a2a]==1.27.2
requests==2.32.5
